+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 - Adds [`Range.prototype.isBounded`][].
 - Adds [`Range.prototype.isUnbounded`][].
 - Adds [`Range.prototype.isFinite`][].
+- Adds [`Range.prototype.isInfinite`][].
 
 [`Range.prototype.isBounded`]: https://github.com/moll/js-strange/blob/master/doc/API.md#Range.prototype.isBounded
 [`Range.prototype.isUnbounded`]: https://github.com/moll/js-strange/blob/master/doc/API.md#Range.prototype.isUnbounded
 [`Range.prototype.isFinite`]: https://github.com/moll/js-strange/blob/master/doc/API.md#Range.prototype.isFinite
+[`Range.prototype.isInfinite`]: https://github.com/moll/js-strange/blob/master/doc/API.md#Range.prototype.isInfinite
 
 ## 1.1.0 (Jun 29, 2015)
 - Adds [`Range.prototype.isEmpty`][].